Move convert to presenter
